+++ FmMain.cs
@@ -682,34 +682,6 @@ namespace TrOCR
 		// - trayExitClick(): 处理“退出”菜单项的点击事件，保存配置、释放资源并终止应用程序。
 		// ====================================================================================================================
 		#region 托盘菜单事件
-		/// <summary>
-		/// 初始化系统托盘菜单项
-		/// 创建包含输入翻译、显示、设置、更新、帮助和退出等功能的托盘右键菜单
-		/// </summary>
-		public void InitMinimize()
-		{
-			try
-			{
-				var menuItems = new[]
-				{
-					new MenuItem("静默识别", traySilentOcrClick),
-					new MenuItem("截图翻译", trayScreenshotTranslateClick),
-					new MenuItem("输入翻译", trayInputTranslateClick),
-					new MenuItem("监听翻译", trayClipListenTranslateClick),
-					new MenuItem("显示", trayShowClick),
-					new MenuItem("设置", tray_Set_Click),
-					new MenuItem("更新", tray_update_Click),
-					new MenuItem("帮助", tray_help_Click),
-					new MenuItem("重启", trayRestartClick), // 【新增】重启菜单项
-					new MenuItem("退出", trayExitClick)
-				};
-				minico.ContextMenu = new ContextMenu(menuItems);
-			}
-			catch (Exception ex)
-			{
-				MessageBox.Show("InitMinimize()" + ex.Message);
-			}
-		}
 		private void btnToggleOriginalText_Click(object sender, EventArgs e)
 		{	
 			if (!(transtalate_fla == "开启"))
@@ -790,143 +762,7 @@ namespace TrOCR
 		        btnToggleOriginalText.Visible = false;
 		    }
 		}
-		/// <summary>
-		/// 托盘菜单"静默识别"选项点击事件处理函数
-		/// </summary>
-		private void traySilentOcrClick(object sender, EventArgs e)
-		{
-            MainSilentOcr();
-            // // 抛出一个带有自定义消息的通用异常
-            // throw new Exception("这是一个测试异常，用于验证 PDB 路径映射是否生效！");
 
-        }
-
-		/// <summary>
-		/// 主静默OCR功能
-		/// </summary>
-		public void MainSilentOcr()
-		{
-		    isSilentMode = true;
-		    MainOCRQuickScreenShots();
-		}
-		private void trayScreenshotTranslateClick(object sender, EventArgs e)
-		{
-		    // 1. 激活截图翻译模式标志
-		    isScreenshotTranslateMode = true;
-
-		    // 2. （可选但推荐）确保其他模式的标志是关闭的，避免冲突
-		    isSilentMode = false;
-
-		    // 3. 调用通用的截图方法，启动流程
-		    MainOCRQuickScreenShots();
-		}
-
-        /// <summary>
-        /// 托盘菜单"输入翻译"选项点击事件处理函数
-        /// 重置翻译界面并显示主输入窗口，根据配置填充剪贴板内容
-        /// </summary>
-        /// <param name="sender">事件发送者</param>
-        /// <param name="e">事件参数</param>
-        private void trayInputTranslateClick(object sender, EventArgs e)
-		{
-			// 标记这不是OCR流程 
-			isContentFromOcr = false;
-			isFromClipboardListener = false;
-
-            // 【新增】重置界面时，确保没有残留的自动翻译任务
-            if (translationTimer != null) translationTimer.Stop();
-
-            // 1. 重置翻译界面，确保只显示主输入窗口			
-            RichBoxBody_T.Visible = false;
-			PictureBox1.Visible = false;
-			RichBoxBody_T.Text = "";
-
-			// 2. 恢复原始窗口大小
-			if (WindowState == FormWindowState.Maximized)
-			{
-				WindowState = FormWindowState.Normal;
-			}
-            transtalate_fla = "关闭";
-			// MinimumSize = new Size((int)font_base.Width * 23, (int)font_base.Height * 24);
-            // Size = new Size((int)font_base.Width * 23, (int)font_base.Height * 24);
-            this.Size = this.lastNormalSize;
-            RichBoxBody.Dock = DockStyle.Fill;          
-
-			// 3. 准备文本内容
-			bool hasContentToTranslate = false;
-			RichBoxBody.richTextBox1.TextChanged -= RichBoxBody_TextChanged; // 关键：在设置文本前，先断开事件处理，避免触发不必要的逻辑
-			try
-			{
-				if (StaticValue.InputTranslateClipboard && Clipboard.ContainsText())
-				{
-					string clipboardText = Clipboard.GetText();
-					string textToDisplay = clipboardText; // 默认显示原始剪贴板文本
-
-					// --- 新增的核心逻辑：检查并执行自动合并 ---
-					if (bool.Parse(IniHelper.GetValue("工具栏", "合并"))) // 检查是否开启了自动合并
-					{
-						if (!string.IsNullOrEmpty(textToDisplay))
-						{
-							string finalText;
-                            if (StaticValue.IsMergeRemoveAllSpace)
-                            {
-                                finalText = Regex.Replace(textToDisplay, @"[\r\n 　]+", "");
-                            }
-							else
-                            {
-                                // 只有在“非移除所有空格”模式下，才调用原来的智能合并方法
-                                finalText = PerformIntelligentMerge(textToDisplay, StaticValue.IsMergeRemoveSpace);
-                            }
-							textToDisplay = finalText;
-
-							// 应用“合并后自动复制”设置
-							if (StaticValue.IsMergeAutoCopy && !string.IsNullOrEmpty(finalText))
-							{
-								SetClipboardWithLock(finalText);
-								Debug.WriteLine("合并后自动复制成功：" + finalText);								
-							}
-						}
-					}
-
-					RichBoxBody.Text = textToDisplay; // 将最终处理好的文本设置到输入框
-
-					if (!string.IsNullOrEmpty(clipboardText))
-					{
-						hasContentToTranslate = true;
-					}
-				}
-				else
-				{
-					RichBoxBody.Text = "";
-				}
-			}
-			finally
-			{
-				RichBoxBody.richTextBox1.TextChanged += RichBoxBody_TextChanged; // 关键：重新订阅事件，以便后续的手动编辑和自动翻译功能正常工作
-			}
-
-			// 4. 显示并激活窗口
-			Show();
-			Activate();
-			Visible = true;
-			WindowState = FormWindowState.Normal;
-			TopMost = IniHelper.GetValue("工具栏", "顶置") == "True";
-
-			// --- 【核心修正】采用更可靠的三步刷新逻辑 ---
-			// 步骤 a: 显式地将焦点设置到文本框，确保它是活动控件
-			RichBoxBody.Focus();
-			// 步骤 b: 处理当前所有Windows消息，确保窗体已完全加载并获得焦点
-			Application.DoEvents();
-			// 步骤 c: 在窗体和控件完全就绪后，再强制刷新，确保渲染正确
-			RichBoxBody.Refresh();
-			// --- 刷新逻辑结束 ---
-
-			// 5. 如果有内容且开启了自动翻译，则手动启动翻译流程
-			if (hasContentToTranslate && StaticValue.InputTranslateAutoTranslate)
-			{
-				TransClick();
-			}
-		}
         /// <summary>
         /// 停止所有正在运行的逻辑定时器 (用于重置状态、关闭窗口或切换模式时)，暂时不使用.手动处理了
 		/// 最核心的原则是：只要发生了“模式切换”或“窗口状态改变”，都应该注意是否停止 translationTimer等定时器。
@@ -942,105 +778,7 @@ namespace TrOCR
             // autoCopyLockTimer 一般不需要强制停止，因为它负责释放锁，强制停止可能导致锁死。
             // trayClickTimer 处理托盘点击，也不建议随意停止。
         }
-        /// <summary>
-        /// 托盘菜单"监听翻译"选项点击事件处理函数
-        /// 用于开启或关闭监听剪贴板进行翻译的功能
-        /// </summary>
-        /// <param name="sender">事件发送者</param>
-        /// <param name="e">事件参数</param>
-        private void trayClipListenTranslateClick(object sender, EventArgs e)
-		{
-			// 1. 切换核心功能状态 (将 StaticValue 中的布尔值取反)
-			StaticValue.ListenClipboardTranslation = !StaticValue.ListenClipboardTranslation;
 
-			// 2. 持久化设置：将新的状态保存到 config.ini 文件
-			IniHelper.SetValue("配置", "ListenClipboard", StaticValue.ListenClipboardTranslation.ToString());
-            // 【新增】如果用户选择关闭监听，立即停止正在等待的防抖定时器
-            if (!StaticValue.ListenClipboardTranslation)
-            {
-                if (clipboardDebounceTimer != null)
-                {
-                    clipboardDebounceTimer.Stop();
-                }
-            }
-
-            // 3. 给用户一个明确的反馈提示
-            if (StaticValue.ListenClipboardTranslation)
-			{
-				CommonHelper.ShowHelpMsg("监听剪贴板翻译已开启");
-			}
-			else
-			{
-				CommonHelper.ShowHelpMsg("监听剪贴板翻译已关闭");
-			}
-		}
-
-		/// <summary>
-		/// 托盘菜单"显示"选项点击事件处理函数
-		/// 显示并激活主窗口
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void trayShowClick(object sender, EventArgs e)
-		{
-			Debug.WriteLine("托盘菜单点击了显示主窗口");
-			Show();
-			Activate();
-			Visible = true;
-			WindowState = FormWindowState.Normal;
-			TopMost = IniHelper.GetValue("工具栏", "顶置") == "True";
-		
-			// 1. 【核心修正】等窗口完全显示、状态稳定了，再绑定事件
-			// 这样可以避开 Show() 过程中可能产生的任何“误触”
-			RichBoxBody.richTextBox1.TextChanged -= RichBoxBody_TextChanged; 
-			RichBoxBody.richTextBox1.TextChanged += RichBoxBody_TextChanged; 
-		}
-
-		/// <summary>
-		/// 托盘菜单"退出"选项点击事件处理函数
-		/// 释放资源并退出应用程序
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void trayExitClick(object sender, EventArgs e)
-		{
-			minico.Dispose();
-			saveIniFile();
-			SaveWindowState();
-			OcrHelper.Dispose();
-			PaddleOCRHelper.Reset();
-			PaddleOCR2Helper.Reset();
-            RapidOCRHelper.Reset();
-            Process.GetCurrentProcess().Kill();
-		}
-		/// <summary>
-		/// 托盘菜单"重启"选项点击事件处理函数
-		/// 保存配置、释放资源并重启应用程序
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void trayRestartClick(object sender, EventArgs e)
-		{
-            // 1. 隐藏并释放托盘图标，防止出现“幽灵图标”
-            minico.Visible = false;
-			minico.Dispose();
-            
-            // 2. 保存当前状态
-			saveIniFile();
-			SaveWindowState();
-            
-            // 3. 释放引擎资源
-			OcrHelper.Dispose();
-			PaddleOCRHelper.Reset();
-			PaddleOCR2Helper.Reset();
-            RapidOCRHelper.Reset();
-
-            // 4. 重启应用程序
-            Application.Restart();
-            
-            // 5. 确保当前进程被完全结束
-			Process.GetCurrentProcess().Kill();
-		}
 		#endregion
 // ====================================================================================================================
 		// **主菜单事件**
@@ -1094,17 +832,12 @@ namespace TrOCR
 		// - OCR_baidu(), OCR_baidu_accurate(): 调用百度标准版和高精度版OCR API。
 		// - OCR_youdao(): 调用有道 OCR API 进行文字识别。
 		// ====================================================================================================================
-		#region OCR 引擎实现
-		#endregion
-
 // ====================================================================================================================
 		// **OCR 接口切换 (事件)**
 		//
 		// 包含用户在界面上选择不同 OCR 引擎的事件处理程序。
 		// 每个事件处理程序通过调用 OCR_foreach(string name) 方法来更新当前使用的 OCR 接口。
 		// ====================================================================================================================
-		#region OCR 接口切换 (事件)
-		#endregion
 // ====================================================================================================================
 		// **文本操作与格式化**
 		//
@@ -1113,10 +846,6 @@ namespace TrOCR
 		// - change_English_Click(): 将文本中的中文标点符号转换为英文标点。
 		// - punctuation_ch_en(): 具体的中文转英文标点实现。
 		// ====================================================================================================================
-		#region 文本操作与格式化
-
-
-		#endregion
 // ====================================================================================================================
 		// **配置文件与初始化**
 		//
@@ -1302,250 +1031,6 @@ namespace TrOCR
 			return new string(array);
 		}
 
-		/// <summary>
-		/// 托盘设置菜单点击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void tray_Set_Click(object sender, EventArgs e)
-		{
-            // 取消注册所有热键
-            HelpWin32.UnregisterHotKey(Handle, 200);
-			HelpWin32.UnregisterHotKey(Handle, 205);
-			HelpWin32.UnregisterHotKey(Handle, 206);
-			HelpWin32.UnregisterHotKey(Handle, 235);
-			HelpWin32.UnregisterHotKey(Handle, 240);
-			HelpWin32.UnregisterHotKey(Handle, 250);
-			HelpWin32.UnregisterHotKey(Handle, 260);
-			
-			WindowState = FormWindowState.Minimized;
-			var fmSetting = new FmSetting();
-			if (settingWindowSize.Width > 574) 
-			{
-                fmSetting.Size = settingWindowSize;
-			}
-			fmSetting.TopMost = true;
-			fmSetting.ShowDialog();
-            //设置窗口关闭后
-            settingWindowSize = fmSetting.Size; // 窗口关闭后，记录它最后的大小
-
-			 //刷新 AI 菜单，这行代码写在fmsetting里也行，写在这里也行
-			LoadCustomOpenAIMenus();
-			LoadCustomOpenAITransMenus();
-			if (fmSetting.DialogResult == DialogResult.OK)
-			{
-				// 在重新加载配置前，保存旧的百度密钥
-				string oldBaiduApiId = StaticValue.BD_API_ID;
-				string oldBaiduApiKey = StaticValue.BD_API_KEY;
-				string oldBaiduAccurateApiId = StaticValue.BD_ACCURATE_API_ID;
-				string oldBaiduAccurateApiKey = StaticValue.BD_ACCURATE_API_KEY;
-
-				var filePath = AppDomain.CurrentDomain.BaseDirectory + "Data\\config.ini";
-				StaticValue.NoteCount = Convert.ToInt32(HelpWin32.IniFileHelper.GetValue("配置", "记录数目", filePath));
-				pubnote = new string[StaticValue.NoteCount];
-				for (var i = 0; i < StaticValue.NoteCount; i++)
-				{
-					pubnote[i] = "";
-				}
-				StaticValue.v_note = pubnote;
-				fmNote.TextNoteChange = "";
-				fmNote.Location = new Point(Screen.AllScreens[0].WorkingArea.Width - fmNote.Width, Screen.AllScreens[0].WorkingArea.Height - fmNote.Height);
-				// 重新注册热键
-				if (IniHelper.GetValue("快捷键", "文字识别") != "请按下快捷键")
-				{
-					var value = IniHelper.GetValue("快捷键", "文字识别");
-					var text = "None";
-					var text2 = "F4";
-					SetHotkey(text, text2, value, 200);
-				}
-				if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-				{
-					var value2 = IniHelper.GetValue("快捷键", "翻译文本");
-					var text3 = "None";
-					var text4 = "F9";
-					SetHotkey(text3, text4, value2, 205);
-				}
-				if (IniHelper.GetValue("快捷键", "记录界面") != "请按下快捷键")
-				{
-					var value3 = IniHelper.GetValue("快捷键", "记录界面");
-					var text5 = "None";
-					var text6 = "F8";
-					SetHotkey(text5, text6, value3, 206);
-				}
-				if (IniHelper.GetValue("快捷键", "识别界面") != "请按下快捷键")
-				{
-					var value4 = IniHelper.GetValue("快捷键", "识别界面");
-					var text7 = "None";
-					var text8 = "F11";
-					SetHotkey(text7, text8, value4, 235);
-				}
-				if (IniHelper.GetValue("快捷键", "输入翻译") != "请按下快捷键")
-				{
-					var value5 = IniHelper.GetValue("快捷键", "输入翻译");
-					// 移除令人困惑的默认键 F1，因为SetHotkey函数会直接解析 value5 字符串
-					SetHotkey("None", "", value5, 240);
-				}
-				if (IniHelper.GetValue("快捷键", "静默识别") != "请按下快捷键")
-				{
-					var value5 = IniHelper.GetValue("快捷键", "静默识别");
-					SetHotkey("None", "", value5, 250);
-				}
-				if (IniHelper.GetValue("快捷键", "截图翻译") != "请按下快捷键")
-				{
-					var value5 = IniHelper.GetValue("快捷键", "截图翻译");
-					SetHotkey("None", "", value5, 260);
-				}
-				// --- 重新加载所有API密钥 ---
-				// --- 加载OCR密钥 ---
-				StaticValue.BD_API_ID = IniHelper.GetValue("密钥_百度", "secret_id");
-				if (StaticValue.BD_API_ID == "发生错误")
-				{
-					StaticValue.BD_API_ID = "";
-				}
-				StaticValue.BD_API_KEY = IniHelper.GetValue("密钥_百度", "secret_key");
-				if (StaticValue.BD_API_KEY == "发生错误")
-				{
-					StaticValue.BD_API_KEY = "";
-				}
-				// 如果百度标准版密钥发生变化，清除旧的Token缓存
-				if (StaticValue.BD_API_ID != oldBaiduApiId || StaticValue.BD_API_KEY != oldBaiduApiKey)
-				{
-					BaiduOcrHelper.ClearAccessTokenCache(false);
-				}
-				StaticValue.BD_LANGUAGE = IniHelper.GetValue("密钥_百度", "language_code");
-				if (StaticValue.BD_LANGUAGE == "发生错误")
-				{
-					StaticValue.BD_LANGUAGE = "CHN_ENG";
-				}
-	
-				StaticValue.TX_API_ID = IniHelper.GetValue("密钥_腾讯", "secret_id");
-				if (StaticValue.TX_API_ID == "发生错误")
-				{
-					StaticValue.TX_API_ID = "";
-				}
-				StaticValue.TX_API_KEY = IniHelper.GetValue("密钥_腾讯", "secret_key");
-				if (StaticValue.TX_API_KEY == "发生错误")
-				{
-					StaticValue.TX_API_KEY = "";
-				}
-				StaticValue.TX_LANGUAGE = IniHelper.GetValue("密钥_腾讯", "language_code");
-				if (StaticValue.TX_LANGUAGE == "发生错误")
-				{
-					StaticValue.TX_LANGUAGE = "zh";
-				}
-	
-				StaticValue.TX_ACCURATE_API_ID = IniHelper.GetValue("密钥_腾讯高精度", "secret_id");
-				if (StaticValue.TX_ACCURATE_API_ID == "发生错误")
-				{
-					StaticValue.TX_ACCURATE_API_ID = "";
-				}
-				StaticValue.TX_ACCURATE_API_KEY = IniHelper.GetValue("密钥_腾讯高精度", "secret_key");
-				if (StaticValue.TX_ACCURATE_API_KEY == "发生错误")
-				{
-					StaticValue.TX_ACCURATE_API_KEY = "";
-				}
-				StaticValue.TX_ACCURATE_LANGUAGE = IniHelper.GetValue("密钥_腾讯高精度", "language");
-				if (StaticValue.TX_ACCURATE_LANGUAGE == "发生错误")
-				{
-					StaticValue.TX_ACCURATE_LANGUAGE = "auto";
-				}
-				StaticValue.TX_TABLE_API_ID = IniHelper.GetValue("密钥_腾讯表格v3", "secret_id");
-				if (StaticValue.TX_TABLE_API_ID == "发生错误")
-				{
-					StaticValue.TX_TABLE_API_ID = "";
-				}
-				StaticValue.TX_TABLE_API_KEY = IniHelper.GetValue("密钥_腾讯表格v3", "secret_key");
-				if (StaticValue.TX_TABLE_API_KEY == "发生错误")
-				{
-					StaticValue.TX_TABLE_API_KEY = "";
-				}
-				
-				StaticValue.BD_ACCURATE_API_ID = IniHelper.GetValue("密钥_百度高精度", "secret_id");
-				if (StaticValue.BD_ACCURATE_API_ID == "发生错误")
-				{
-				    StaticValue.BD_ACCURATE_API_ID = "";
-				}
-				StaticValue.BD_ACCURATE_API_KEY = IniHelper.GetValue("密钥_百度高精度", "secret_key");
-				if (StaticValue.BD_ACCURATE_API_KEY == "发生错误")
-				{
-				    StaticValue.BD_ACCURATE_API_KEY = "";
-				}
-				// 如果百度高精度版密钥发生变化，清除旧的Token缓存
-				if (StaticValue.BD_ACCURATE_API_ID != oldBaiduAccurateApiId || StaticValue.BD_ACCURATE_API_KEY != oldBaiduAccurateApiKey)
-				{
-					BaiduOcrHelper.ClearAccessTokenCache(true);
-				}
-				StaticValue.BD_ACCURATE_LANGUAGE = IniHelper.GetValue("密钥_百度高精度", "language_code");
-				if (StaticValue.BD_ACCURATE_LANGUAGE == "发生错误")
-				{
-				    StaticValue.BD_ACCURATE_LANGUAGE = "CHN_ENG";
-				}
-				// 重新加载百度表格识别密钥
-				StaticValue.BD_TABLE_API_ID = IniHelper.GetValue("密钥_百度表格", "secret_id");
-				if (StaticValue.BD_TABLE_API_ID == "发生错误")
-				{
-					StaticValue.BD_TABLE_API_ID = "";
-				}
-				StaticValue.BD_TABLE_API_KEY = IniHelper.GetValue("密钥_百度表格", "secret_key");
-				if (StaticValue.BD_TABLE_API_KEY == "发生错误")
-				{
-					StaticValue.BD_TABLE_API_KEY = "";
-				}
-				// 【新增】重新加载百度手写识别密钥
-				StaticValue.BD_HANDWRITING_API_ID = TrOCRUtils.LoadSetting("密钥_百度手写", "secret_id", "");
-				StaticValue.BD_HANDWRITING_API_KEY = TrOCRUtils.LoadSetting("密钥_百度手写", "secret_key", "");
-				StaticValue.BD_HANDWRITING_LANGUAGE = TrOCRUtils.LoadSetting("密钥_百度手写", "language_code", "CHN_ENG");
-	
-				// --- 重新加载白描OCR凭据 ---
-				string newBaimiaoUsername = IniHelper.GetValue("密钥_白描", "username");
-				if (newBaimiaoUsername == "发生错误") newBaimiaoUsername = "";
-
-				// 如果用户名发生变化，则清除旧的token缓存
-				if (StaticValue.BaimiaoUsername != newBaimiaoUsername)
-				{
-				 OcrHelper.ClearBaimiaoTokenCache();
-				}
-				StaticValue.BaimiaoUsername = newBaimiaoUsername;
-
-				StaticValue.BaimiaoPassword = IniHelper.GetValue("密钥_白描", "password");
-				if (StaticValue.BaimiaoPassword == "发生错误") StaticValue.BaimiaoPassword = "";
-
-				// 重新加载持久化的token信息
-				string savedToken = IniHelper.GetValue("密钥_白描", "token");
-				string savedExpiry = IniHelper.GetValue("密钥_白描", "token_expiry");
-				string savedUsername = IniHelper.GetValue("密钥_白描", "token_username");
-				if (!string.IsNullOrEmpty(savedToken) && savedToken != "发生错误" &&
-				    !string.IsNullOrEmpty(savedUsername) && savedUsername != "发生错误" &&
-				    savedUsername == StaticValue.BaimiaoUsername &&
-				    DateTime.TryParse(savedExpiry, out DateTime expiry) && DateTime.Now < expiry)
-				{
-				 StaticValue.BaimiaoToken = savedToken;
-				 StaticValue.BaimiaoTokenExpiry = expiry;
-				}
-				else
-				{
-				 StaticValue.BaimiaoToken = null;
-				 StaticValue.BaimiaoTokenExpiry = DateTime.MinValue;
-				}
-
-				// 重新加载UUID
-				string savedUuid = IniHelper.GetValue("密钥_白描", "device_uuid");
-				StaticValue.BaimiaoDeviceUuid = (savedUuid == "发生错误" || string.IsNullOrEmpty(savedUuid)) ? null : savedUuid;
-
-				// 重新加载翻译配置
-				StaticValue.Translate_Current_API = IniHelper.GetValue("配置", "翻译接口");
-				if (StaticValue.Translate_Current_API == "发生错误")
-				{
-					StaticValue.Translate_Current_API = "Bing2";
-				}
-				LoadTranslateConfig();
-				InitializeApiMenus();
-				  // --- 【关键新增】刷新工具栏图标状态 ---
-        		this.RichBoxBody.readIniFile();
-        		this.RichBoxBody_T.readIniFile();
-			}
-		}
-
 
         #endregion
 
@@ -1561,30 +1046,6 @@ namespace TrOCR
         // ====================================================================================================================
         #region 翻译功能
 
-		/// <summary>
-		/// 动态计算加载图标(PictureBox1)的位置，使其在翻译文本框(RichBoxBody_T)中居中显示。
-		/// </summary>
-		private void PositionLoadingIcon()
-		{
-			// 确保翻译框是可见的，否则无法定位
-			if (!RichBoxBody_T.Visible)
-			{
-				RichBoxBody_T.Visible = true;
-			}
-		
-			// 1. 计算图标应该在的X, Y坐标
-			//    公式: 容器的起始位置 + (容器宽度 - 图标宽度) / 2
-			int centerX = RichBoxBody_T.Left + (RichBoxBody_T.Width - PictureBox1.Width) / 2;
-			int centerY = RichBoxBody_T.Top + (RichBoxBody_T.Height - PictureBox1.Height) / 2;
-		
-			// 2. 应用新的坐标
-			//    使用 Math.Max 确保图标不会因窗口过小而跑到负坐标
-			PictureBox1.Location = new Point(Math.Max(0, centerX), Math.Max(0, centerY));
-		
-			// 3. 确保图标可见并位于最顶层
-			PictureBox1.Visible = true;
-			PictureBox1.BringToFront();
-		}
 		/// <summary>
 		/// 处理窗体大小调整事件，当翻译功能开启时调整文本框大小和位置
 		/// </summary>
@@ -1934,157 +1395,9 @@ namespace TrOCR
 		    }
 		    
 		}
-		/// <summary>
-		/// 显示加载窗口并运行应用程序消息循环
-		/// </summary>
-		private void ShowLoading()
-		{
-			try
-			{
-				fmloading = new FmLoading();
-				Application.Run(fmloading);
-			}
-			catch (ThreadAbortException)
-			{
-			}
-			catch (Exception ex)
-			{
-				MessageBox.Show(ex.ToString());
-			}
-			finally
-			{
-				thread.Abort();
-			}
-		}
-
 
 #endregion
 
-// ====================================================================================================================
-		// **文本朗读 (TTS)**
-		//
-		// 实现了文本到语音的转换功能。
-		// - TTS(): 启动一个新的线程来处理 TTS 请求。
-		// - TTS_thread(): 在后台线程中获取要朗读的文本，检测语言，并从百度 TTS 服务下载音频数据。
-		// - TTS_child(): 在 UI 线程中播放下载的音频。
-		// ====================================================================================================================
-#region 文本朗读 (TTS)
-		/// <summary>
-		/// 启动TTS文本朗读功能，在新线程中执行TTS_thread方法
-		/// </summary>
-		public void TTS()
-		{
-			new Thread(TTS_thread).Start();
-		}
-
-		/// <summary>
-		/// TTS文本朗读线程函数，负责获取文本内容、检测语言、下载语音数据并调用播放方法
-		/// </summary>
-		public void TTS_thread()
-		{
-            try
-            {
-                // 1. === 获取 Token (使用 using 自动释放) ===
-                string text;
-                HttpWebRequest tokenRequest = (HttpWebRequest)WebRequest.Create(string.Format("{0}?{1}", "http://aip.baidubce.com/oauth/2.0/token", "grant_type=client_credentials&client_id=iQekhH39WqHoxur5ss59GpU4&client_secret=8bcee1cee76ed60cdfaed1f2c038584d"));
-
-                using (HttpWebResponse tokenResponse = (HttpWebResponse)tokenRequest.GetResponse())
-                using (Stream responseStream = tokenResponse.GetResponseStream())
-                using (StreamReader streamReader = new StreamReader(responseStream, Encoding.GetEncoding("utf-8")))
-                {
-                    text = streamReader.ReadToEnd();
-                }
-                // <--- 在这里，tokenResponse, responseStream, 和 streamReader 已被自动关闭和释放
-
-                string text2 = !contain_ch(htmltxt) ? "zh" : "zh";
-
-                // 2. === 构建 TTS 请求 ===
-                HttpWebRequest httpWebRequest = (HttpWebRequest)WebRequest.Create(string.Concat(new string[]
-                {
-            		"http://tsn.baidu.com/text2audio?lan=" + text2 + "&ctp=1&cuid=abcdxxx&tok=",
-            		((JObject)JsonConvert.DeserializeObject(text))["access_token"].ToString(),
-            		"&tex=",
-            		HttpUtility.UrlEncode(htmltxt.Replace("***", "")),
-            		"&vol=9&per=0&spd=5&pit=5"
-                }));
-                httpWebRequest.Method = "POST";
-
-                byte[] array2;
-
-                // 3. === 获取音频流 (使用 using 自动释放) ===
-                // 修复了原版代码在 while 循环中重复调用 GetResponseStream() 的 Bug
-                using (HttpWebResponse httpWebResponse = (HttpWebResponse)httpWebRequest.GetResponse())
-                using (Stream audioStream = httpWebResponse.GetResponseStream()) // <-- 只调用一次
-                using (MemoryStream memoryStream = new MemoryStream())
-                {
-                    byte[] array = new byte[16384];
-                    int num;
-                    while ((num = audioStream.Read(array, 0, array.Length)) > 0)
-                    {
-                        memoryStream.Write(array, 0, num);
-                    }
-                    array2 = memoryStream.ToArray();
-                }
-                // <--- 在这里，httpWebResponse, audioStream, 和 memoryStream 已被自动关闭和释放
-
-                ttsData = array2;
-                if (speak_copyb == "朗读" || voice_count == 0)
-                {
-                    Invoke(new Translate(Speak_child));
-                    speak_copyb = "";
-                }
-                else
-                {
-                    Invoke(new Translate(TTS_child));
-                }
-                voice_count++;
-            }
-            catch (Exception ex)
-            {
-                if (ex.ToString().IndexOf("Null") <= -1)
-                {
-                    MessageBox.Show("文本过长，请使用右键菜单中的选中朗读！", "提醒");
-                }
-            }
-        }
-
-		/// <summary>
-		/// TTS文本朗读播放函数，在UI线程中执行，负责播放已下载的语音数据
-		/// </summary>
-		public void TTS_child()
-		{
-			// 检查主文本框或翻译文本框是否有内容
-			if (RichBoxBody.Text != null || RichBoxBody_T.Text != "")
-			{
-				// 如果正在播放，则关闭播放并返回
-				if (speaking)
-				{
-					HelpWin32.mciSendString("close media", null, 0, IntPtr.Zero);
-					speaking = false;
-					return;
-				}
-				// 获取系统临时目录路径
-				var tempPath = Path.GetTempPath();
-				// 构造临时音频文件路径
-				var text = tempPath + "\\声音.mp3";
-				try
-				{
-					// 将语音数据写入临时文件
-					File.WriteAllBytes(text, ttsData);
-				}
-				catch
-				{
-					// 如果写入失败，尝试使用另一个文件名
-					text = tempPath + "\\声音1.mp3";
-					File.WriteAllBytes(text, ttsData);
-				}
-				// 播放音频文件
-				PlaySong(text);
-				// 设置播放状态为正在播放
-				speaking = true;
-			}
-		}
-#endregion
 // ====================================================================================================================
 		// **截图与图像处理**
 		//
@@ -2144,16 +1457,6 @@ namespace TrOCR
 			Process.Start("https://www.baidu.com/s?wd=" + RichBoxBody.SelectText);
 		}
 
-		/// <summary>
-		/// 点击托盘更新菜单项时检查程序更新
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void tray_update_Click(object sender, EventArgs e)
-		{
-			Program.CheckUpdate();
-		}
-
 
 		/// <summary>
 		/// 重命名文件，在文件名后添加序号以避免重复
@@ -2176,16 +1479,6 @@ namespace TrOCR
 			return Path.GetFileName(path);
 		}
 
-		/// <summary>
-		/// 播放指定的音频文件
-		/// </summary>
-		/// <param name="file">音频文件路径</param>
-		public void PlaySong(string file)
-		{
-			HelpWin32.mciSendString("close media", null, 0, IntPtr.Zero);
-			HelpWin32.mciSendString("open \"" + file + "\" type mpegvideo alias media", null, 0, IntPtr.Zero);
-			HelpWin32.mciSendString("play media notify", null, 0, Handle);
-		}
 #endregion
 
 // ====================================================================================================================
@@ -2196,60 +1489,6 @@ namespace TrOCR
 		// - Trans_Voice_Click(): 获取译文框中选中的文本并触发朗读。
 		// - Speak_child(): 在 UI 线程中播放朗读音频。
 		// ====================================================================================================================
-#region 右键菜单 - 朗读事件
-		/// <summary>
-		/// 主文本框语音朗读点击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void Main_Voice_Click(object sender, EventArgs e)
-		{
-			RichBoxBody.Focus();
-			speak_copyb = "朗读";
-			htmltxt = RichBoxBody.SelectText;
-			HelpWin32.SendMessage(Handle, 786, 590);
-		}
-
-		/// <summary>
-		/// 翻译文本框语音朗读点击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void Trans_Voice_Click(object sender, EventArgs e)
-		{
-			RichBoxBody_T.Focus();
-			speak_copyb = "朗读";
-			htmltxt = RichBoxBody_T.SelectText;
-			HelpWin32.SendMessage(Handle, 786, 590);
-		}
-
-		/// <summary>
-		/// 执行文本语音朗读功能
-		/// </summary>
-		public void Speak_child()
-		{
-			// 检查主文本框或翻译文本框是否有内容
-			if (RichBoxBody.Text != null || RichBoxBody_T.Text != "")
-			{
-				var tempPath = Path.GetTempPath();
-				var text = tempPath + "\\声音.mp3";
-				try
-				{
-					File.WriteAllBytes(text, ttsData);
-				}
-				catch
-				{
-					text = tempPath + "\\声音1.mp3";
-					File.WriteAllBytes(text, ttsData);
-				}
-				PlaySong(text);
-				speaking = true;
-			}
-		}
-
-
-#endregion
-
 // ====================================================================================================================
 		// **右键菜单 - 文本转换**
 		//
@@ -2259,11 +1498,6 @@ namespace TrOCR
 		// - change_str_Upper_Click(): 将文本转换为大写。
 		// - change_Upper_str_Click(): 将文本转换为小写。
 		// ====================================================================================================================
-#region 右键菜单 - 文本转换
-
-
-#endregion
-
 // ====================================================================================================================
 		// **热键管理**
 		//
@@ -2392,1108 +1626,6 @@ namespace TrOCR
 
 
 		/// <summary>
-		/// 托盘帮助菜单点击事件处理函数
-		/// 最小化当前窗口并打开帮助窗口
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void tray_help_Click(object sender, EventArgs e)
-		{
-			WindowState = FormWindowState.Minimized;
-			new FmHelp().Show();
-		}
-
-
-		/// <summary>
-		/// 在输入图像中查找轮廓并为每个轮廓绘制边界框，将结果绘制到目标图像上
-		/// </summary>
-		/// <param name="src">输入的灰度图像，用于查找轮廓</param>
-		/// <param name="draw">用于绘制结果的目标图像</param>
-		/// <returns>带有边界框的图像</returns>
-		public Image BoundingBox(Image<Gray, byte> src, Image<Bgr, byte> draw)
-		{
-			Image result;
-			using (var vectorOfVectorOfPoint = new VectorOfVectorOfPoint())
-			{
-				// 查找图像中的轮廓
-				CvInvoke.FindContours(src, vectorOfVectorOfPoint, null, RetrType.List, ChainApproxMethod.ChainApproxSimple);
-				Image image = draw.ToBitmap();
-				var graphics = Graphics.FromImage(image);
-				var size = vectorOfVectorOfPoint.Size;
-				// 遍历所有轮廓并绘制边界框
-				for (var i = 0; i < size; i++)
-				{
-					using (var vectorOfPoint = vectorOfVectorOfPoint[i])
-					{
-						var rectangle = CvInvoke.BoundingRectangle(vectorOfPoint);
-						var x = rectangle.Location.X;
-						var y = rectangle.Location.Y;
-						var width = rectangle.Size.Width;
-						var height = rectangle.Size.Height;
-						// 只处理大于5x5像素的轮廓
-						if (width > 5 || height > 5)
-						{
-							graphics.FillRectangle(Brushes.White, x, 0, width, image.Size.Height);
-						}
-					}
-				}
-				graphics.Dispose();
-				// 创建一个稍大的新位图以容纳结果
-				var bitmap = new Bitmap(image.Width + 2, image.Height + 2);
-				var graphics2 = Graphics.FromImage(bitmap);
-				graphics2.DrawImage(image, 1, 1, image.Width, image.Height);
-				graphics2.Save();
-				graphics2.Dispose();
-				result = bitmap;
-			}
-			return result;
-		}
-
-		/// <summary>
-		/// 从源图像中查找轮廓并提取感兴趣的区域图像
-		/// </summary>
-		/// <param name="src">输入的灰度图像，用于查找轮廓</param>
-		/// <param name="draw">输出图像，用于绘制结果</param>
-		public void select_image(Image<Gray, byte> src, Image<Bgr, byte> draw)
-		{
-			try
-			{
-				// 查找图像中的轮廓
-				using (var vectorOfVectorOfPoint = new VectorOfVectorOfPoint())
-				{
-					CvInvoke.FindContours(src, vectorOfVectorOfPoint, null, RetrType.List, ChainApproxMethod.ChainApproxSimple);
-					var num = vectorOfVectorOfPoint.Size / 2;
-					imagelist_lenght = num;
-					bool_image_count(num);
-					
-					// 确保临时图像目录存在
-					if (!Directory.Exists(AppDomain.CurrentDomain.BaseDirectory + "Data\\image_temp"))
-					{
-						Directory.CreateDirectory(AppDomain.CurrentDomain.BaseDirectory + "Data\\image_temp");
-					}
-					
-					// 清空OCR结果变量
-					OCR_baidu_a = "";
-					OCR_baidu_b = "";
-					OCR_baidu_c = "";
-					OCR_baidu_d = "";
-					OCR_baidu_e = "";
-					
-					// 遍历所有轮廓，提取对应的图像区域
-					for (var i = 0; i < num; i++)
-					{
-						using (var vectorOfPoint = vectorOfVectorOfPoint[i])
-						{
-							var rectangle = CvInvoke.BoundingRectangle(vectorOfPoint);
-							if (rectangle.Size.Width > 1 && rectangle.Size.Height > 1)
-							{
-								var x = rectangle.Location.X;
-								var y = rectangle.Location.Y;
-								var width = rectangle.Size.Width;
-								var height = rectangle.Size.Height;
-								new Point(x, 0);
-								new Point(x, image_ori.Size.Height);
-								var srcRect = new Rectangle(x, 0, width, image_ori.Size.Height);
-								var bitmap = new Bitmap(width + 70, srcRect.Size.Height);
-								var graphics = Graphics.FromImage(bitmap);
-								graphics.FillRectangle(Brushes.White, 0, 0, bitmap.Size.Width, bitmap.Size.Height);
-								graphics.DrawImage(image_ori, 30, 0, srcRect, GraphicsUnit.Pixel);
-								var bitmap2 = Image.FromHbitmap(bitmap.GetHbitmap());
-								bitmap2.Save("Data\\image_temp\\" + i + ".jpg", ImageFormat.Jpeg);
-								bitmap2.Dispose();
-								bitmap.Dispose();
-								graphics.Dispose();
-							}
-						}
-					}
-					
-					// 显示加载消息对话框
-					var messageload = new Messageload();
-					messageload.ShowDialog();
-					if (messageload.DialogResult == DialogResult.OK)
-					{
-						// 启动后台工作线程
-						var array = new[]
-						{
-							new ManualResetEvent(false)
-						};
-						ThreadPool.QueueUserWorkItem(DoWork, array[0]);
-					}
-				}
-			}
-			catch
-			{
-				exit_thread();
-			}
-		}
-
-		/// <summary>
-		/// 查找图像中的边界框
-		/// 该函数使用OpenCV处理图像，通过灰度化、腐蚀、阈值处理和边缘检测等步骤，
-		/// 最终识别出图像中的主要对象并绘制边界框
-		/// </summary>
-		/// <param name="bitmap">需要处理的原始图像</param>
-		/// <returns>带有边界框标记的图像</returns>
-		public Image FindBundingBox(Bitmap bitmap)
-		{
-            var image = bitmap.ToImage<Bgr, byte>();
-            var image2 = new Image<Gray, byte>(image.Width, image.Height);
-			CvInvoke.CvtColor(image, image2, ColorConversion.Bgra2Gray);
-			var structuringElement = CvInvoke.GetStructuringElement(ElementShape.Rectangle, new Size(4, 4), new Point(1, 1));
-			CvInvoke.Erode(image2, image2, structuringElement, new Point(0, 2), 1, BorderType.Reflect101, default(MCvScalar));
-			CvInvoke.Threshold(image2, image2, 100.0, 255.0, (ThresholdType)9);
-            var image3 = image2.ToBitmap().ToImage<Gray, byte>();
-            var draw = image3.Convert<Bgr, byte>();
-			var image4 = image3.Clone();
-			CvInvoke.Canny(image3, image4, 255.0, 255.0, 5, true);
-			return BoundingBox(image4, draw);
-		}
-
-
-		/// <summary>
-		/// 对指定范围的图像文件进行OCR识别（处理第一部分图像）
-		/// </summary>
-		/// <param name="objEvent">线程同步事件对象</param>
-		public void baidu_image_a(object objEvent)
-		{
-			try
-			{
-				// 批量处理第一部分图片
-				for (var i = 0; i < image_num[0]; i++)
-				{
-					Stream stream = File.Open("Data\\image_temp\\" + i + ".jpg", FileMode.Open);
-					OcrBdUseA(Image.FromStream(stream));
-					stream.Close();
-				}
-				((ManualResetEvent)objEvent).Set();
-			}
-			catch
-			{
-				exit_thread();
-			}
-		}
-
-		/// <summary>
-		/// 对指定范围的图像文件进行OCR识别（处理第二部分图像）
-		/// </summary>
-		/// <param name="objEvent">线程同步事件对象</param>
-		public void baidu_image_b(object objEvent)
-		{
-			try
-			{
-				// 批量处理第二部分图片
-				for (var i = image_num[0]; i < image_num[1]; i++)
-				{
-					Stream stream = File.Open("Data\\image_temp\\" + i + ".jpg", FileMode.Open);
-					OcrBdUseB(Image.FromStream(stream));
-					stream.Close();
-				}
-				((ManualResetEvent)objEvent).Set();
-			}
-			catch
-			{
-				exit_thread();
-			}
-		}
-
-		private void DoWork(object state)
-		{
-			/// <summary>
-			/// 执行OCR识别工作，处理竖排文字识别任务
-			/// 启动多个线程分别处理不同区域的图片OCR识别，等待所有识别完成后整合结果
-			/// </summary>
-			/// <param name="state">线程状态参数</param>
-			
-			// 创建5个ManualResetEvent用于线程同步
-			var array = new ManualResetEvent[5];
-			array[0] = new ManualResetEvent(false);
-			// 启动线程处理第一部分图片OCR识别
-			ThreadPool.QueueUserWorkItem(baidu_image_a, array[0]);
-			array[1] = new ManualResetEvent(false);
-			// 启动线程处理第二部分图片OCR识别
-			ThreadPool.QueueUserWorkItem(baidu_image_b, array[1]);
-			array[2] = new ManualResetEvent(false);
-			// 启动线程处理第三部分图片OCR识别
-			ThreadPool.QueueUserWorkItem(BdImageC, array[2]);
-			array[3] = new ManualResetEvent(false);
-			// 启动线程处理第四部分图片OCR识别
-			ThreadPool.QueueUserWorkItem(BdImageD, array[3]);
-			array[4] = new ManualResetEvent(false);
-			// 启动线程处理第五部分图片OCR识别
-			ThreadPool.QueueUserWorkItem(BdImageE, array[4]);
-			WaitHandle[] waitHandles = array;
-			// 等待所有OCR识别线程完成
-			WaitHandle.WaitAll(waitHandles);
-			// 整合所有OCR识别结果并去除多余换行符
-			shupai_Right_txt = string.Concat(OCR_baidu_a, OCR_baidu_b, OCR_baidu_c, OCR_baidu_d, OCR_baidu_e).Replace("\r\n\r\n", "");
-			var text = shupai_Right_txt.TrimEnd('\n').TrimEnd('\r').TrimEnd('\n');
-			// 如果识别结果包含多行文本，则进行文本方向调整
-			if (text.Split(Environment.NewLine.ToCharArray()).Length > 1)
-			{
-				var array2 = text.Split(new[]
-				{
-					"\r\n"
-				}, StringSplitOptions.None);
-				var str = "";
-				// 反转文本行顺序以适应从右到左的阅读顺序
-				for (var i = 0; i < array2.Length; i++)
-				{
-					str = str + array2[array2.Length - i - 1].Replace("\r", "").Replace("\n", "") + "\r\n";
-				}
-				shupai_Left_txt = str;
-			}
-			fmloading.FmlClose = "窗体已关闭";
-			Invoke(new OcrThread(Main_OCR_Thread_last));
-			try
-			{
-				// 清理临时图片文件
-				DeleteFile("Data\\image_temp");
-			}
-			catch
-			{
-				exit_thread();
-			}
-			// 释放原图资源
-			image_ori.Dispose();
-		}
-
-		/// <summary>
-		/// 删除指定路径的文件或目录
-		/// </summary>
-		/// <param name="path">要删除的文件或目录路径</param>
-		public void DeleteFile(string path)
-		{
-			if (File.GetAttributes(path) == FileAttributes.Directory)
-			{
-				Directory.Delete(path, true);
-				return;
-			}
-			File.Delete(path);
-		}
-
-		/// <summary>
-		/// 处理image_num[1]到image_num[2]范围内的图片文件，使用OcrBdUseC进行OCR识别
-		/// </summary>
-		/// <param name="objEvent">用于线程同步的ManualResetEvent对象</param>
-		public void BdImageC(object objEvent)
-		{
-			try
-			{
-				for (var i = image_num[1]; i < image_num[2]; i++)
-				{
-					Stream stream = File.Open("Data\\image_temp\\" + i + ".jpg", FileMode.Open);
-					OcrBdUseC(Image.FromStream(stream));
-					stream.Close();
-				}
-				((ManualResetEvent)objEvent).Set();
-			}
-			catch
-			{
-				exit_thread();
-			}
-		}
-
-		/// <summary>
-		/// 处理image_num[2]到image_num[3]范围内的图片文件，使用OcrBdUseD进行OCR识别
-		/// </summary>
-		/// <param name="objEvent">用于线程同步的ManualResetEvent对象</param>
-		public void BdImageD(object objEvent)
-		{
-			try
-			{
-				for (var i = image_num[2]; i < image_num[3]; i++)
-				{
-					Stream stream = File.Open("Data\\image_temp\\" + i + ".jpg", FileMode.Open);
-					OcrBdUseD(Image.FromStream(stream));
-					stream.Close();
-				}
-				((ManualResetEvent)objEvent).Set();
-			}
-			catch
-			{
-				exit_thread();
-			}
-		}
-
-		/// <summary>
-		/// 处理image_num[3]到image_num[4]范围内的图片文件，使用OcrBdUseE进行OCR识别
-		/// </summary>
-		/// <param name="objEvent">用于线程同步的ManualResetEvent对象</param>
-		public void BdImageE(object objEvent)
-		{
-			try
-			{
-				for (var i = image_num[3]; i < image_num[4]; i++)
-				{
-					Stream stream = File.Open("Data\\image_temp\\" + i + ".jpg", FileMode.Open);
-					OcrBdUseE(Image.FromStream(stream));
-					stream.Close();
-				}
-				((ManualResetEvent)objEvent).Set();
-			}
-			catch
-			{
-				exit_thread();
-			}
-		}
-
-		/// <summary>
-		/// 根据传入的数字参数，计算并设置image_num数组的值。
-		/// 该函数主要用于将输入的数字分成5个区间，每个区间包含相对均匀的数量。
-		/// </summary>
-		/// <param name="num">需要处理的总数</param>
-		public void bool_image_count(int num)
-		{
-			// 当数量大于等于5时，将数据分为5个区间
-			if (num >= 5)
-			{
-				image_num = new int[num];
-				// 根据余数的不同情况，分别计算各区间边界值
-				if (num - num / 5 * 5 == 0)
-				{
-					image_num[0] = num / 5;
-					image_num[1] = num / 5 * 2;
-					image_num[2] = num / 5 * 3;
-					image_num[3] = num / 5 * 4;
-					image_num[4] = num;
-				}
-				if (num - num / 5 * 5 == 1)
-				{
-					image_num[0] = num / 5 + 1;
-					image_num[1] = num / 5 * 2;
-					image_num[2] = num / 5 * 3;
-					image_num[3] = num / 5 * 4;
-					image_num[4] = num;
-				}
-				if (num - num / 5 * 5 == 2)
-				{
-					image_num[0] = num / 5 + 1;
-					image_num[1] = num / 5 * 2 + 1;
-					image_num[2] = num / 5 * 3;
-					image_num[3] = num / 5 * 4;
-					image_num[4] = num;
-				}
-				if (num - num / 5 * 5 == 3)
-				{
-					image_num[0] = num / 5 + 1;
-					image_num[1] = num / 5 * 2 + 1;
-					image_num[2] = num / 5 * 3 + 1;
-					image_num[3] = num / 5 * 4;
-					image_num[4] = num;
-				}
-				if (num - num / 5 * 5 == 4)
-				{
-					image_num[0] = num / 5 + 1;
-					image_num[1] = num / 5 * 2 + 1;
-					image_num[2] = num / 5 * 3 + 1;
-					image_num[3] = num / 5 * 4 + 1;
-					image_num[4] = num;
-				}
-			}
-			// 处理数量为4的特殊情况
-			if (num == 4)
-			{
-				image_num = new int[5];
-				image_num[0] = 1;
-				image_num[1] = 2;
-				image_num[2] = 3;
-				image_num[3] = 4;
-				image_num[4] = 0;
-			}
-			// 处理数量为3的特殊情况
-			if (num == 3)
-			{
-				image_num = new int[5];
-				image_num[0] = 1;
-				image_num[1] = 2;
-				image_num[2] = 3;
-				image_num[3] = 0;
-				image_num[4] = 0;
-			}
-			// 处理数量为2的特殊情况
-			if (num == 2)
-			{
-				image_num = new int[5];
-				image_num[0] = 1;
-				image_num[1] = 2;
-				image_num[2] = 0;
-				image_num[3] = 0;
-				image_num[4] = 0;
-			}
-			// 处理数量为1的特殊情况
-			if (num == 1)
-			{
-				image_num = new int[5];
-				image_num[0] = 1;
-				image_num[1] = 0;
-				image_num[2] = 0;
-				image_num[3] = 0;
-				image_num[4] = 0;
-			}
-			// 处理数量为0的特殊情况
-			if (num == 0)
-			{
-				image_num = new int[5];
-				image_num[0] = 0;
-				image_num[1] = 0;
-				image_num[2] = 0;
-				image_num[3] = 0;
-				image_num[4] = 0;
-			}
-		}
-
-		/// <summary>
-		/// 退出线程处理方法，用于停止当前的截图线程并恢复窗体状态
-		/// </summary>
-		private void exit_thread()
-		{
-			try
-			{
-				// 停止截图操作
-				StaticValue.IsCapture = false;
-				esc = "退出";
-				// 关闭加载窗体
-				fmloading.FmlClose = "窗体已关闭";
-				// 终止截图线程
-				esc_thread.Abort();
-			}
-			catch
-			{
-				//
-			}
-			// 恢复主窗体状态
-			FormBorderStyle = FormBorderStyle.Sizable;
-			Visible = true;
-			Show();
-			WindowState = FormWindowState.Normal;
-			// 重新设置翻译文本的快捷键
-			if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-			{
-				var value = IniHelper.GetValue("快捷键", "翻译文本");
-				var text = "None";
-				var text2 = "F9";
-				SetHotkey(text, text2, value, 205);
-			}
-			// 注销热键
-			HelpWin32.UnregisterHotKey(Handle, 222);
-		}
-
-
-		/// <summary>
-		/// 缩放图像到指定尺寸
-		/// </summary>
-		/// <param name="bitmap1">需要缩放的原始图像</param>
-		/// <param name="destHeight">目标最小高度</param>
-		/// <param name="destWidth">目标最小宽度</param>
-		/// <returns>缩放后的图像</returns>
-		private Bitmap ZoomImage(Bitmap bitmap1, int destHeight, int destWidth)
-		{
-			// 获取原始图像的宽度和高度
-			var num = (double)bitmap1.Width;
-			var num2 = (double)bitmap1.Height;
-			// 如果宽度小于目标高度，则等比例放大
-			if (num < destHeight)
-			{
-				while (num < destHeight)
-				{
-					num2 *= 1.1;
-					num *= 1.1;
-				}
-			}
-			// 如果高度小于目标宽度，则等比例放大
-			if (num2 < destWidth)
-			{
-				while (num2 < destWidth)
-				{
-					num2 *= 1.1;
-					num *= 1.1;
-				}
-			}
-			// 转换为整数尺寸
-			var width = (int)num;
-			var height = (int)num2;
-			// 创建新图像并绘制缩放后的图像
-			var bitmap2 = new Bitmap(width, height);
-			var graphics = Graphics.FromImage(bitmap2);
-			graphics.DrawImage(bitmap1, 0, 0, width, height);
-			graphics.Save();
-			graphics.Dispose();
-			return new Bitmap(bitmap2);
-		}
-
-
-		/// <summary>
-		/// 从指定图像中提取矩形区域并返回新的位图
-		/// </summary>
-		/// <param name="pic">源图像</param>
-		/// <param name="rect">要提取的矩形区域</param>
-		/// <returns>提取出的矩形区域位图</returns>
-		public Bitmap GetRect(Image pic, Rectangle rect)
-		{
-			var destRect = new Rectangle(0, 0, rect.Width, rect.Height);
-			var bitmap = new Bitmap(destRect.Width, destRect.Height);
-			var graphics = Graphics.FromImage(bitmap);
-			graphics.Clear(Color.FromArgb(0, 0, 0, 0));
-			graphics.DrawImage(pic, destRect, rect, GraphicsUnit.Pixel);
-			graphics.Dispose();
-			return bitmap;
-		}
-
-		/// <summary>
-		/// 从给定图像中提取指定区域的子图像，并保存为PNG文件
-		/// </summary>
-		/// <param name="buildPic">源图像，用于提取子图像</param>
-		/// <param name="buildRects">矩形区域数组，指定要从源图像中提取的区域</param>
-		/// <returns>提取出的子图像数组</returns>
-		private Bitmap[] getSubPics(Image buildPic, Rectangle[] buildRects)
-		{
-			var array = new Bitmap[buildRects.Length];
-			for (var i = 0; i < buildRects.Length; i++)
-			{
-				array[i] = GetRect(buildPic, buildRects[i]);
-				var filename = IniHelper.GetValue("配置", "截图位置") + "\\" + ReFileName(IniHelper.GetValue("配置", "截图位置"), "图片.Png");
-				array[i].Save(filename, ImageFormat.Png);
-			}
-			return array;
-		}
-
-		/// <summary>
-		/// 从给定图像中提取指定区域的子图像，并对每个子图像执行OCR识别
-		/// </summary>
-		/// <param name="buildPic">源图像，用于提取子图像</param>
-		/// <param name="buildRects">矩形区域数组，指定要从源图像中提取的区域</param>
-		/// <returns>提取出的子图像数组</returns>
-		private Bitmap[] getSubPics_ocr(Image buildPic, Rectangle[] buildRects)
-		{
-			var text = "";
-			var array = new Bitmap[buildRects.Length];
-			var text2 = "";
-			for (var i = 0; i < buildRects.Length; i++)
-			{
-				// 提取指定区域的子图像
-				array[i] = GetRect(buildPic, buildRects[i]);
-				image_screen = array[i];
-				var messageload = new Messageload();
-				messageload.ShowDialog();
-				if (messageload.DialogResult == DialogResult.OK)
-				{
-					// 根据选择的OCR接口执行相应的OCR识别方法
-					if (interface_flag == "搜狗")
-					{
-						SougouOCR();
-					}
-					if (interface_flag == "腾讯")
-					{
-						OCR_Tencent();
-					}
-					if (interface_flag == "有道")
-					{
-						OCR_youdao();
-					}
-					if (interface_flag == "白描")
-					{
-						OCR_Baimiao();
-					}
-					if (interface_flag == "日语" || interface_flag == "中英" || interface_flag == "韩语")
-					{
-						OCR_baidu();
-					}
-					messageload.Dispose();
-				}
-				// 根据配置和段落标志处理识别结果文本
-				if (IniHelper.GetValue("工具栏", "分栏") == "True")
-				{
-					if (paragraph)
-					{
-						text = text + "\r\n" + typeset_txt.Trim();
-						text2 = text2 + "\r\n" + split_txt.Trim() + "\r\n";
-					}
-					else
-					{
-						text += typeset_txt.Trim();
-						text2 = text2 + "\r\n" + split_txt.Trim() + "\r\n";
-					}
-				}
-				else if (paragraph)
-				{
-					text = text + "\r\n" + typeset_txt.Trim() + "\r\n";
-					text2 = text2 + "\r\n" + split_txt.Trim() + "\r\n";
-				}
-				else
-				{
-					text = text + typeset_txt.Trim() + "\r\n";
-					text2 = text2 + "\r\n" + split_txt.Trim() + "\r\n";
-				}
-			}
-			// 整理识别结果，去除多余的换行符
-			typeset_txt = text.Replace("\r\n\r\n", "\r\n");
-			split_txt = text2.Replace("\r\n\r\n", "\r\n");
-			fmloading.FmlClose = "窗体已关闭";
-			Invoke(new OcrThread(Main_OCR_Thread_last));
-			return array;
-		}
-
-		/// <summary>
-		/// 查找图像中围栏区域的边界框并进行处理
-		/// 该函数使用OpenCV库对输入图像进行处理，识别围栏状结构，提取对应的边界框区域用于后续OCR识别
-		/// </summary>
-		/// <param name="src">输入的灰度图像，用于查找轮廓</param>
-		/// <param name="draw">用于绘制结果的彩色图像</param>
-		/// <returns>处理后的图像，其中围栏区域被标记</returns>
-		public Image BoundingBox_fences(Image<Gray, byte> src, Image<Bgr, byte> draw)
-		{
-			Image result;
-			using (var vectorOfVectorOfPoint = new VectorOfVectorOfPoint())
-			{
-				CvInvoke.FindContours(src, vectorOfVectorOfPoint, null, RetrType.List, ChainApproxMethod.ChainApproxSimple);
-				Image image = draw.ToBitmap();
-				var graphics = Graphics.FromImage(image);
-				var size = vectorOfVectorOfPoint.Size;
-				for (var i = 0; i < size; i++)
-				{
-					using (var vectorOfPoint = vectorOfVectorOfPoint[i])
-					{
-						var rectangle = CvInvoke.BoundingRectangle(vectorOfPoint);
-						var x = rectangle.Location.X;
-						var y = rectangle.Location.Y;
-						var width = rectangle.Size.Width;
-						var height = rectangle.Size.Height;
-						graphics.FillRectangle(Brushes.White, x, 0, width, draw.Height);
-					}
-				}
-				graphics.Dispose();
-				var bitmap = new Bitmap(image.Width + 2, image.Height + 2);
-				var graphics2 = Graphics.FromImage(bitmap);
-				graphics2.DrawImage(image, 1, 1, image.Width, image.Height);
-				graphics2.Save();
-				graphics2.Dispose();
-				image.Dispose();
-				src.Dispose();
-				result = bitmap;
-			}
-			return result;
-		}
-
-		/// <summary>
-		/// 查找图像中围栏区域的边界框并进行处理
-		/// 该函数使用OpenCV库对输入图像进行处理，识别围栏状结构，提取对应的边界框区域用于后续OCR识别
-		/// </summary>
-		/// <param name="bitmap">输入的位图图像</param>
-		/// <returns>处理后的图像，其中围栏区域被标记</returns>
-		public Image FindBoundingBoxFences(Bitmap bitmap)
-		{
-			var image = bitmap.ToImage<Bgr, byte>();
-			var image2 = new Image<Gray, byte>(image.Width, image.Height);
-			// 将彩色图像转换为灰度图像
-			CvInvoke.CvtColor(image, image2, ColorConversion.Bgra2Gray);
-			// 创建结构元素用于形态学操作
-			var structuringElement = CvInvoke.GetStructuringElement(ElementShape.Rectangle, new Size(6, 20), new Point(1, 1));
-			// 对图像进行腐蚀操作，增强围栏特征
-			CvInvoke.Erode(image2, image2, structuringElement, new Point(0, 2), 1, BorderType.Reflect101, default(MCvScalar));
-			// 应用阈值处理将图像二值化
-			CvInvoke.Threshold(image2, image2, 100.0, 255.0, (ThresholdType)9);
-			var image3 = image2.ToBitmap().ToImage<Gray, byte>();
-			var draw = image3.Convert<Bgr, byte>();
-			// 复制图像用于边缘检测
-			var image4 = image3.Clone();
-			// 使用Canny算法检测图像边缘
-			CvInvoke.Canny(image3, image4, 255.0, 255.0, 5, true);
-			// 查找并标记边界框区域
-			var image5 = BoundingBox_fences(image4, draw);
-            var image6 = ((Bitmap)image5).ToImage<Gray, byte>();
-            // 对标记的区域进行进一步处理
-            BoundingBox_fences_Up(image6);
-			// 释放资源
-			image.Dispose();
-			image2.Dispose();
-			image3.Dispose();
-			image6.Dispose();
-			return image5;
-		}
-
-		/// <summary>
-		/// 查找图像中的轮廓并提取对应的边界框区域用于OCR识别
-		/// 该函数使用OpenCV库来查找图像中的轮廓，并为每个轮廓创建边界矩形，然后调用OCR处理函数
-		/// </summary>
-		/// <param name="src">输入的灰度图像，用于查找轮廓</param>
-		public void BoundingBox_fences_Up(Image<Gray, byte> src)
-		{
-			using (var vectorOfVectorOfPoint = new VectorOfVectorOfPoint())
-			{
-				// 查找图像中的所有轮廓
-				CvInvoke.FindContours(src, vectorOfVectorOfPoint, null, RetrType.List, ChainApproxMethod.ChainApproxSimple);
-				var size = vectorOfVectorOfPoint.Size;
-				// 为每个轮廓创建对应的边界矩形
-				var array = new Rectangle[size];
-				// 遍历所有轮廓，获取边界矩形并按相反顺序存储
-				for (var i = 0; i < size; i++)
-				{
-					using (var vectorOfPoint = vectorOfVectorOfPoint[i])
-					{
-						array[size - 1 - i] = CvInvoke.BoundingRectangle(vectorOfPoint);
-					}
-				}
-				// 对提取的子图像区域进行OCR识别处理
-				getSubPics_ocr(image_screen, array);
-			}
-		}
-
-
-		/// <summary>
-		/// 托盘图标双击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		/// <summary>
-		/// 【新增】托盘图标鼠标按下事件
-		/// </summary>
-		private void tray_MouseDown(object sender, MouseEventArgs e)
-		{
-			if (e.Button == MouseButtons.Left)
-			{
-				// 每当左键按下，计数器加 1
-				trayClickCount++;
-
-				if (trayClickCount == 1)
-				{
-					// 如果这是第一次点击，启动定时器
-					// 定时器将等待系统双击那么长的时间
-					trayClickTimer.Start();
-				}
-			}
-		}
-	/// <summary>
-    /// 【修改】定时器触发事件：检查时间窗口内的总点击次数
-    /// </summary>
-    private void trayClickTimer_Tick(object sender, EventArgs e)
-    {
-        // 1. 定时器到期，立即停止
-        trayClickTimer.Stop();
-
-        // 2. 检查在时间窗口内总共发生了几次点击
-        if (trayClickCount == 1)
-        {
-            // 如果只有 1 次点击，判定为“单击”，执行“切换显示/隐藏”
-				if (this.Visible)
-				{ 	
-					// 隐藏：先解绑
-            		RichBoxBody.richTextBox1.TextChanged -= RichBoxBody_TextChanged;
-					 // 【新增】防止后台读秒
-            		if (translationTimer != null) translationTimer.Stop(); 
-					// 如果窗口当前可见，则隐藏它
-					this.Hide();
-					this.Visible = false;
-				}
-				else
-				{
-					// 如果窗口当前不可见，则调用 trayShowClick (它会负责 Show, Activate, TopMost 等)
-					trayShowClick(sender, e);
-				}
-        }
-        else if (trayClickCount >= 2)
-        {
-            // 如果有 2 次或更多点击，判定为“双击”
-            // (这里我们执行原 tray_double_Click 的逻辑)
-            HelpWin32.UnregisterHotKey(Handle, 205);
-            menu.Hide();
-            RichBoxBody.Hide = "";
-            RichBoxBody_T.Hide = "";
-            MainOCRQuickScreenShots();
-        }
-
-        // 3. 无论结果如何，重置计数器
-        trayClickCount = 0;
-    }
-
-
-		private void ExportExcel_Click(object sender, EventArgs e)
-		{
-		    // 调用我们创建的帮助类来处理导出
-		    // Helper.ExcelHelper.ExportToExcel(this.lastRecognizedTable,this.lastRecognizedHeader, this.lastRecognizedFooter);
-			if (this.lastOcrProvider == "Baidu")
-		    {
-		        Helper.ExcelHelper.ExportToExcel(this.lastBaiduCells, this.lastRecognizedHeader, this.lastRecognizedFooter,this);
-		    }
-		    else if (this.lastOcrProvider == "Tencent")
-		    {
-		        Helper.ExcelHelper.ExportToExcel(this.lastTencentCells, this.lastRecognizedHeader, this.lastRecognizedFooter,this);
-		    }
-		    else
-		    {
-		        MessageBox.Show("没有可供导出的表格数据。", "信息", MessageBoxButtons.OK, MessageBoxIcon.Information);
-		    }
-		}
-
-		/// <summary>
-		/// 使用百度API进行表格OCR识别
-		/// 该方法会截取当前屏幕图像，调用百度表格OCR API进行识别，并将结果处理后显示在RichBoxBody中
-		/// </summary>
-		public void BdTableOCR()
-		{
-			typeset_txt = "[消息]：表格已下载！";
-			split_txt = "";
-			// this.lastRecognizedTable = null; // 每次识别前清空 
-    		this.lastRecognizedHeader = null;
-    		this.lastRecognizedFooter = null;
-			this.lastBaiduCells = null;
-    		this.lastOcrProvider = "Baidu";
-			try
-			{
-				// 获取图像字节数组
-				var image = image_screen;
-				var imageBytes = OcrHelper.ImgToBytes(image);
-
-				// 调用新的表格识别方法
-				// string result = BaiduOcrHelper.TableRecognition(imageBytes, out this.lastRecognizedTable, out this.lastRecognizedHeader, out this.lastRecognizedFooter, false, false);
-				DataTable dummyTable; // DataTable 现在只是个附属品
-				string result = BaiduOcrHelper.TableRecognition(imageBytes, out dummyTable, out this.lastRecognizedHeader, out this.lastRecognizedFooter, out this.lastBaiduCells, false, false);
-
-				// 检查识别结果
-				if (string.IsNullOrWhiteSpace(result))
-				{
-					typeset_txt = "***该区域未发现文本***";
-				}
-				else
-				{
-					// 设置识别结果
-					typeset_txt = result;
-				}
-				split_txt = "";
-			}
-			catch (Exception ex)
-			{
-				typeset_txt = $"[消息]：表格识别异常: {ex.Message}";
-			}
-		}
-
-		/// <summary>
-		/// 腾讯表格OCR识别方法
-		/// 该方法会截取当前屏幕图像，调用腾讯表格OCR API进行识别，并将结果处理后显示在RichBoxBody中
-		/// </summary>
-		public void TxTableOCR()
-		{
-			typeset_txt = "[消息]：表格已下载！";
-			split_txt = "";
-			// this.lastRecognizedTable = null; 
-    		this.lastRecognizedHeader = null;
-    		this.lastRecognizedFooter = null;
-			this.lastTencentCells = null;
-    		this.lastOcrProvider = "Tencent";
-			try
-			{
-				// 获取图像字节数组
-				var image = image_screen;
-				var imageBytes = OcrHelper.ImgToBytes(image);
-
-				// 调用腾讯表格识别方法
-				// string result = TencentOcrHelper.TableRecognition(imageBytes,out this.lastRecognizedTable,out this.lastRecognizedHeader, out this.lastRecognizedFooter);
-				DataTable dummyTable;
-				string result = TencentOcrHelper.TableRecognition(imageBytes, out dummyTable, out this.lastRecognizedHeader, out this.lastRecognizedFooter, out this.lastTencentCells);
-				// 检查识别结果
-				if (string.IsNullOrWhiteSpace(result))
-				{
-					typeset_txt = "***该区域未发现文本***";
-				}
-				else
-				{
-					// 设置识别结果
-					typeset_txt = result;
-				}
-				split_txt = "";
-			}
-			catch (Exception ex)
-			{
-				typeset_txt = $"[消息]：表格识别异常: {ex.Message}";
-			}
-		}
-
-		/// <summary>
-		/// 表格OCR按钮点击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void OCR_table_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("表格");
-		}
-
-		/// <summary>
-		/// 表格OCR主线程处理函数
-		/// 该方法处理OCR识别完成后的结果展示和相关清理工作
-		/// </summary>
-		public void Main_OCR_Thread_table()
-		{
-			ailibaba = new AliTable();
-			var timeSpan = new TimeSpan(DateTime.Now.Ticks);
-			var timeSpan2 = timeSpan.Subtract(ts).Duration();
-			var str = string.Concat(new[]
-			{
-				timeSpan2.Seconds.ToString(),
-				".",
-				Convert.ToInt32(timeSpan2.TotalMilliseconds).ToString(),
-				"秒"
-			});
-			// 根据配置设置窗口是否置顶
-			if (StaticValue.v_topmost)
-			{
-				TopMost = true;
-			}
-			else
-			{
-				TopMost = false;
-			}
-			Text = "耗时：" + str;
-			// 根据接口类型处理识别结果
-			if (interface_flag == "百度表格")
-			{
-				// 1. 检查识别是否成功
-				bool isSuccess = !string.IsNullOrWhiteSpace(typeset_txt) &&
-								 !typeset_txt.Contains("***请在设置中输入百度标准版密钥或表格识别专用密钥***") &&
-								 !typeset_txt.Contains("***该区域未发现文本***") &&
-								 !typeset_txt.Contains("***该区域未发现表格***") &&
-								 !typeset_txt.Contains("错误") &&
-								 !typeset_txt.Contains("异常") &&
-								 !typeset_txt.Contains("失败");
-
-				// 如果识别成功，复制到剪贴板
-				if (isSuccess)
-				{
-					// 显示识别结果
-					RichBoxBody.Text = "[消息]：百度表格识别成功，已复制到粘贴板！可直接导出Excel文件或粘贴到Excel";
-
-					// 提取HTML表格部分
-					string htmlTable = ExtractHtmlTable(typeset_txt);
-
-					if (!string.IsNullOrEmpty(htmlTable))
-					{
-						// 设置HTML格式到剪贴板，Excel可以识别并保持表格结构
-						var dataObject = new DataObject();
-						dataObject.SetData(DataFormats.Html, CreateHtmlClipboardData(htmlTable));
-						dataObject.SetData(DataFormats.UnicodeText, typeset_txt);
-						SetClipboardWithLock(dataObject);
-						Debug.WriteLine("百度识别表格结果复制到剪贴板-html格式");
-					}
-					else
-					{
-						// 如果没有HTML表格，使用普通文本
-						var dataObject = new DataObject();
-						dataObject.SetData(DataFormats.UnicodeText, typeset_txt);
-						SetClipboardWithLock(dataObject);
-						Debug.WriteLine("百度识别表格结果复制到剪贴板-普通文本格式");
-					}
-
-				}
-				else
-				{
-					// 如果失败，显示实际的错误信息
-					RichBoxBody.Text = typeset_txt;
-				}
-			}
-			else if (interface_flag == "腾讯表格")
-			{
-				// 检查腾讯表格识别是否成功
-				bool isSuccess = !string.IsNullOrWhiteSpace(typeset_txt) &&
-								 !typeset_txt.Contains("***请在设置中输入腾讯标准版密钥或表格识别专用密钥***") &&
-								 !typeset_txt.Contains("***该区域未发现文本***") &&
-								 !typeset_txt.Contains("***该区域未发现表格***") &&
-								 !typeset_txt.Contains("错误") &&
-								 !typeset_txt.Contains("异常") &&
-								 !typeset_txt.Contains("失败");
-
-				// 如果识别成功，复制到剪贴板
-				if (isSuccess)
-				{
-					// 显示识别结果
-					RichBoxBody.Text = "[消息]：腾讯表格识别成功，已复制到粘贴板！可直接导出Excel文件或粘贴到Excel";
-
-					// 提取HTML表格部分
-					string htmlTable = ExtractHtmlTable(typeset_txt);
-
-					if (!string.IsNullOrEmpty(htmlTable))
-					{
-						// 设置HTML格式到剪贴板，Excel可以识别并保持表格结构
-						var dataObject = new DataObject();
-						dataObject.SetData(DataFormats.Html, CreateHtmlClipboardData(htmlTable));
-						dataObject.SetData(DataFormats.UnicodeText, typeset_txt);
-						SetClipboardWithLock(dataObject);
-						Debug.WriteLine("腾讯识别表格结果复制到剪贴板-html格式");
-					}
-					else
-					{
-						// 如果没有HTML表格，使用普通文本
-						var dataObject = new DataObject();
-						dataObject.SetData(DataFormats.UnicodeText, typeset_txt);
-						SetClipboardWithLock(dataObject);
-						Debug.WriteLine("腾讯识别表格结果复制到剪贴板-普通文本格式");
-					}
-				}
-				else
-				{
-					// 如果失败，显示实际的错误信息
-					RichBoxBody.Text = typeset_txt;
-				}
-			}
-			
-			// 清理资源
-			image_screen.Dispose();
-			GC.Collect();
-			StaticValue.IsCapture = false;
-			FormBorderStyle = FormBorderStyle.Sizable;
-			Visible = true;
-			Show();
-			WindowState = FormWindowState.Normal;
-			// Size = new Size(form_width, form_height);
-			this.Size = this.lastNormalSize;
-			HelpWin32.SetForegroundWindow(Handle);
-			if (interface_flag == "阿里表格")
-			{
-				if (split_txt == "弹出cookie")
-				{
-					split_txt = "";
-					ailibaba.TopMost = true;
-					ailibaba.getcookie = "";
-					IniHelper.SetValue("特殊", "ali_cookie", ailibaba.getcookie);
-					ailibaba.ShowDialog();
-					HelpWin32.SetForegroundWindow(ailibaba.Handle);
-					return;
-				}
-				SetClipboardWithLock(typeset_txt);
-				Debug.WriteLine("阿里表格结果剪切板写入成功");
-				CopyHtmlToClipBoard(typeset_txt);
-			}
-			
-		}
-
-		/// <summary>
-		/// 百度表格OCR识别按钮点击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void OCR_baidutable_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("百度表格");
-		}
-		/// <summary>
-		/// 腾讯表格OCR识别按钮点击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void OCR_txtable_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("腾讯表格");
-		}
-
-		/// <summary>
-		/// 阿里表格OCR识别按钮点击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void OCR_ailitable_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("阿里表格");
-		}
-
-		/// <summary>
 		/// 刷新ocr接口界面控件文本显示
 		/// </summary>
 		private new void Refresh()
@@ -3526,104 +1658,6 @@ namespace TrOCR
 
 		}
 
-		/// <summary>
-		/// 将Image对象转换为字节数组
-		/// </summary>
-		/// <param name="img">要转换的Image对象</param>
-		/// <returns>表示图像数据的字节数组</returns>
-		public static byte[] ImageToByteArray(Image img)
-		{
-			return (byte[])new ImageConverter().ConvertTo(img, typeof(byte[]));
-		}
-
-		/// <summary>
-		/// 将字节数组转换为Stream对象
-		/// </summary>
-		/// <param name="bytes">要转换的字节数组</param>
-		/// <returns>包含字节数据的Stream对象</returns>
-		public static Stream BytesToStream(byte[] bytes)
-		{
-			return new MemoryStream(bytes);
-		}
-
-		/// <summary>
-		/// 使用阿里云OCR服务识别表格
-		/// </summary>
-		public void OCR_ali_table()
-		{
-			var text = "";
-			split_txt = "";
-			try
-			{
-				var value = IniHelper.GetValue("特殊", "ali_cookie");
-				var stream = BytesToStream(ImageToByteArray(BWPic((Bitmap)image_screen)));
-				var str = Convert.ToBase64String(new BinaryReader(stream).ReadBytes(Convert.ToInt32(stream.Length)));
-				stream.Close();
-				var postStr = "{\n\t\"image\": \"" + str + "\",\n\t\"configure\": \"{\\\"format\\\":\\\"html\\\", \\\"finance\\\":false}\"\n}";
-				var url = "https://predict-pai.data.aliyun.com/dp_experience_mall/ocr/ocr_table_parse";
-				text = CommonHelper.PostStrData(url, postStr, value);
-				typeset_txt = ((JObject)JsonConvert.DeserializeObject(CommonHelper.PostStrData(url, postStr, value)))["tables"].ToString().Replace("table tr td { border: 1px solid blue }", "table tr td {border: 0.5px black solid }").Replace("table { border: 1px solid blue }", "table { border: 0.5px black solid; border-collapse : collapse}\r\n");
-				RichBoxBody.Text = "[消息]：表格已复制到粘贴板！";
-			}
-			catch
-			{
-				RichBoxBody.Text = "[消息]：阿里表格识别出错！";
-				if (text.Contains("NEED_LOGIN"))
-				{
-					split_txt = "弹出cookie";
-				}
-			}
-		}
-
-		/// <summary>
-		/// 将彩色图像转换为黑白图像
-		/// </summary>
-		/// <param name="mybm">需要转换的原始彩色图像</param>
-		/// <returns>转换后的黑白图像</returns>
-		public Bitmap BWPic(Bitmap mybm)
-		{
-			var bitmap = new Bitmap(mybm.Width, mybm.Height);
-			// 遍历图像中的每个像素点
-			for (var i = 0; i < mybm.Width; i++)
-			{
-				for (var j = 0; j < mybm.Height; j++)
-				{
-					var pixel = mybm.GetPixel(i, j);
-					// 通过计算RGB三个分量的平均值来获得灰度值
-					var num = (pixel.R + pixel.G + pixel.B) / 3;
-					bitmap.SetPixel(i, j, Color.FromArgb(num, num, num));
-				}
-			}
-			return bitmap;
-		}
-
-		/// <summary>
-		/// 将HTML内容复制到剪贴板
-		/// </summary>
-		/// <param name="html">要复制到剪贴板的HTML内容</param>
-		public void CopyHtmlToClipBoard(string html)
-		{
-			var utf = Encoding.UTF8;
-			// HTML剪贴板格式的标准头部信息
-			var format = "Version:0.9\r\nStartHTML:{0:000000}\r\nEndHTML:{1:000000}\r\nStartFragment:{2:000000}\r\nEndFragment:{3:000000}\r\n";
-			// HTML片段的开始标记和结束标记
-			var text = "<html>\r\n<head>\r\n<meta http-equiv=\"Content-Type\" content=\"text/html; charset=" + utf.WebName + "\">\r\n<title>HTML clipboard</title>\r\n</head>\r\n<body>\r\n<!--StartFragment-->";
-			var text2 = "<!--EndFragment-->\r\n</body>\r\n</html>\r\n";
-			// 计算各个部分的字节位置
-			var s = string.Format(format, 0, 0, 0, 0);
-			var byteCount = utf.GetByteCount(s);
-			var byteCount2 = utf.GetByteCount(text);
-			var byteCount3 = utf.GetByteCount(html);
-			var byteCount4 = utf.GetByteCount(text2);
-			// 构造完整的HTML剪贴板数据
-			var s2 = string.Format(format, byteCount, byteCount + byteCount2 + byteCount3 + byteCount4, byteCount + byteCount2, byteCount + byteCount2 + byteCount3) + text + html + text2;
-			var dataObject = new DataObject();
-			dataObject.SetData(DataFormats.Html, new MemoryStream(utf.GetBytes(s2)));
-			var data = new HtmlToText().Convert(html);
-			dataObject.SetData(DataFormats.UnicodeText, data);
-			SetClipboardWithLock(dataObject);
-			Debug.WriteLine("识别表格结果写入剪贴板");
-		}
 
 #endregion
 		//控制右键菜单-接口项的可见性
@@ -4271,68 +2305,6 @@ namespace TrOCR
 				sb.Append(b.ToString("x2"));
 			// 返回加密的字符串
 			return sb.ToString();
-		}
-
-		/// <summary>
-		/// 从文本中提取HTML表格部分
-		/// </summary>
-		/// <param name="text">包含HTML表格的文本</param>
-		/// <returns>HTML表格字符串，如果没有找到则返回空字符串</returns>
-		private string ExtractHtmlTable(string text)
-		{
-			if (string.IsNullOrEmpty(text))
-				return string.Empty;
-
-			int startIndex = text.IndexOf("<table");
-			if (startIndex == -1)
-				return string.Empty;
-
-			int endIndex = text.IndexOf("</table>", startIndex);
-			if (endIndex == -1)
-				return string.Empty;
-
-			endIndex += "</table>".Length;
-			return text.Substring(startIndex, endIndex - startIndex);
-		}
-
-		/// <summary>
-		/// 创建HTML剪贴板数据格式
-		/// </summary>
-		/// <param name="htmlTable">HTML表格字符串</param>
-		/// <returns>符合剪贴板格式的HTML数据</returns>
-		private string CreateHtmlClipboardData(string htmlTable)
-		{
-			if (string.IsNullOrEmpty(htmlTable))
-				return string.Empty;
-
-			// HTML剪贴板格式需要特定的头部信息
-			string htmlClipboardData = 
-				"Version:0.9" + Environment.NewLine +
-				"StartHTML:0000000000" + Environment.NewLine +
-				"EndHTML:0000000000" + Environment.NewLine +
-				"StartFragment:0000000000" + Environment.NewLine +
-				"EndFragment:0000000000" + Environment.NewLine +
-				"<html>" + Environment.NewLine +
-				"<body>" + Environment.NewLine +
-				"<!--StartFragment-->" + Environment.NewLine +
-				htmlTable + Environment.NewLine +
-				"<!--EndFragment-->" + Environment.NewLine +
-				"</body>" + Environment.NewLine +
-				"</html>";
-
-			// 计算偏移量
-			int startHTML = htmlClipboardData.IndexOf("<html>");
-			int endHTML = htmlClipboardData.IndexOf("</html>") + "</html>".Length;
-			int startFragment = htmlClipboardData.IndexOf("<!--StartFragment-->") + "<!--StartFragment-->".Length;
-			int endFragment = htmlClipboardData.IndexOf("<!--EndFragment-->");
-
-			// 更新偏移量（10位数字，左补0）
-			htmlClipboardData = htmlClipboardData.Replace("StartHTML:0000000000", string.Format("StartHTML:{0:D10}", startHTML));
-			htmlClipboardData = htmlClipboardData.Replace("EndHTML:0000000000", string.Format("EndHTML:{0:D10}", endHTML));
-			htmlClipboardData = htmlClipboardData.Replace("StartFragment:0000000000", string.Format("StartFragment:{0:D10}", startFragment));
-			htmlClipboardData = htmlClipboardData.Replace("EndFragment:0000000000", string.Format("EndFragment:{0:D10}", endFragment));
-
-			return htmlClipboardData;
 		}
